parser: Define ASTs in separate module
